requirements.txt
asgiref==3.8.1
black==24.8.0
click==8.1.7
crispy-bootstrap5==2024.2
dj-database-url==2.2.0
Django==5.1
django-crispy-forms==2.3
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.3
python-dateutil==2.9.0.post0
python-decouple==3.8
six==1.16.0
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
